+++ flutter/integration_test/utils.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 
 import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/ui/root/main_screen.dart';
-import 'package:mlperfbench/ui/run/result_screen.dart';
+import 'package:mlperfbench/ui/home/result_screen.dart';
 import 'package:mlperfbench_common/data/extended_result.dart';
 import 'package:mlperfbench/resources/result_manager.dart' as result_manager;
 import 'package:mlperfbench/resources/resource_manager.dart'
