+++ flutter/lib/ui/home/progress_screen.dart
@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
-import 'package:mlperfbench/ui/run/progress_circles.dart';
+import 'package:mlperfbench/ui/home/progress_circles.dart';
 import 'package:mlperfbench/ui/time_utils.dart';
 
 class ProgressScreen extends StatefulWidget {
