+++ flutter/lib/ui/root/main_screen.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:mlperfbench/benchmark/state.dart';
+import 'package:mlperfbench/ui/home/progress_screen.dart';
+import 'package:mlperfbench/ui/home/result_screen.dart';
+import 'package:mlperfbench/ui/home/start_screen.dart';
 import 'package:mlperfbench/ui/root/resource_error_screen.dart';
-import 'package:mlperfbench/ui/run/progress_screen.dart';
-import 'package:mlperfbench/ui/run/result_screen.dart';
-import 'package:mlperfbench/ui/run/start_screen.dart';
 
-class MyHome extends StatelessWidget {
-  const MyHome({Key? key}) : super(key: key);
+class MainScreen extends StatelessWidget {
+  const MainScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
